CMP-171: Changed the data type for "amount" and "captured amount" from a 32-bit integer to a 64-bit long (which in PHP is handled as a float). This has affected the following methods: - formatAmount from class: General - valAmount from class: Validate - valPrice from class: Validate Additionally class: TxnInfo now handles both "amount" and "captured amount" as floats to provide support for numbers larger than: 2.147.483.647. This has also affected the following columns in the data model: - amount from table: Log.Transaction_Tbl - captured from table: Log.Transaction_Tbl Finally any type casting of the "amount" has been changed to use PHP's "floatval" function in the following controllers: - /mApp/api/initialize.php - /anet/dpm.php
git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@809 86c063f6-31ef-4d33-b9ba-d2d3e4924f03

+++ webroot/mApp/api/initialize.php
@@ -86,7 +86,7 @@ if (array_key_exists("PHP_AUTH_USER", $_SERVER) === true && array_key_exists("PH
 						$obj_mPoint->newMessage($iTxnID, Constants::iINPUT_VALID_STATE, $obj_DOM->asXML() );
 
 						$data['typeid'] = $obj_DOM->{'initialize-payment'}[$i]->transaction["type-id"];
-						$data['amount'] = (integer) $obj_DOM->{'initialize-payment'}[$i]->transaction->amount;
+						$data['amount'] = (float) $obj_DOM->{'initialize-payment'}[$i]->transaction->amount;
 						$data['country-config'] = $obj_CountryConfig;
 						if (count($obj_DOM->{'initialize-payment'}[$i]->transaction->points) == 1)
 						{
